+++ src/Codenizer.HttpClient.Testable/RequestMethodNode.cs
@@ -29,7 +29,16 @@ namespace Codenizer.HttpClient.Testable
 
         public RequestSchemeNode? Match(string scheme)
         {
-            return _schemeNodes.SingleOrDefault(s => s.Scheme == scheme);
+            var explicitSchemeMatch = _schemeNodes.SingleOrDefault(s => s.Scheme == scheme);
+
+            if (explicitSchemeMatch == null)
+            {
+                // If no explicit match was found, try to use a wildcard match.
+                // This is applicable for responses with a relative URI.
+                return _schemeNodes.SingleOrDefault(s => s.Scheme == "*");
+            }
+
+            return explicitSchemeMatch;
         }
     }
 }