+++ src/Codenizer.HttpClient.Testable/RequestPathNode.cs
@@ -13,13 +13,15 @@ namespace Codenizer.HttpClient.Testable
             Path = path;
         }
 
-        public RequestQueryNode Add(List<KeyValuePair<string, string>> queryParameters)
+        public RequestQueryNode Add(
+            List<KeyValuePair<string, string>> queryParameters,
+            List<QueryStringAssertion> queryStringAssertions)
         {
             var existingQuery = _queryNodes.SingleOrDefault(node => node.Matches(queryParameters));
 
             if (existingQuery == null)
             {
-                existingQuery = new RequestQueryNode(queryParameters);
+                existingQuery = new RequestQueryNode(queryParameters, queryStringAssertions);
                 _queryNodes.Add(existingQuery);
             }
 
@@ -29,6 +31,51 @@ namespace Codenizer.HttpClient.Testable
         public RequestQueryNode? Match(string? queryString)
         {
             return _queryNodes.SingleOrDefault(node => node.Matches(queryString));
+        }
+
+        public bool MatchesPath(string path)
+        {
+            if (PathHasRouteParameters())
+            {
+                return MatchesPathWithRouteParameters(path);
+            }
+
+            return Path == path;
+        }
+
+        private bool PathHasRouteParameters()
+        {
+            return Path.Contains("{") && Path.Contains("}");
+        }
+
+        private bool MatchesPathWithRouteParameters(string path)
+        {
+            var pathSegments = Path.Split('/');
+            var matchSegments = path.Split('/');
+
+            if (pathSegments.Length != matchSegments.Length)
+            {
+                return false;
+            }
+
+            for (var index = 0; index < pathSegments.Length; index++)
+            {
+                // Check whether this segment is a route parameter (for example: '{id}')
+                if (pathSegments[index].StartsWith("{") && pathSegments[index].EndsWith("}"))
+                {
+                    // It is, ignore this particular segment for matching
+                    // because we treat this as a wildcard.
+                    continue;
+                }
+
+                if (pathSegments[index] != matchSegments[index])
+                {
+
+                    return false;
+                }
+            }
+
+            return true;
         }
     }
 }